view 제작 준비

+++ src/main/java/com/kgd/springbootmall/entity/seller/Color.java
@@ -18,7 +18,7 @@ public class Color {
     @Id
     @Column(nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Double num;
+    private Long num;
 
     private String name;
